+++ lumen/.gitignore
@@ -1,4 +1,5 @@
 /vendor
+/node_modules
 .env
 .env.production
 .env.local
