+++ lumen/package.json
@@ -1,0 +1,8 @@
+{
+  "private": true,
+  "devDependencies": {
+    "gulp": "^3.8.8",
+    "gulp-minify-css": "^1.2.2",
+    "gulp-uglify": "^1.5.1"
+  }
+}
