--- public_html/css/buttons.css
@@ -1,106 +1,0 @@
-.button,
-button,
-input[type="submit"],
-input[type="reset"],
-input[type="button"] {
-    display: inline-block;
-    height: 38px;
-    padding: 0 30px;
-    color: #555;
-    text-align: center;
-    font-size: 11px;
-    font-weight: 600;
-    line-height: 38px;
-    letter-spacing: .1rem;
-    text-transform: uppercase;
-    text-decoration: none;
-    white-space: nowrap;
-    background-color: transparent;
-    border-radius: 4px;
-    border: 1px solid #bbb;
-    cursor: pointer;
-    box-sizing: border-box; }
-.button:hover,
-button:hover,
-input[type="submit"]:hover,
-input[type="reset"]:hover,
-input[type="button"]:hover,
-.button:focus,
-button:focus,
-input[type="submit"]:focus,
-input[type="reset"]:focus,
-input[type="button"]:focus {
-    color: #333;
-    border-color: #888;
-    outline: 0; }
-
-.button.button-primary,
-button.button-primary,
-input[type="submit"].button-primary,
-input[type="reset"].button-primary,
-input[type="button"].button-primary {
-    color: #FFF;
-    background-color: #33C3F0;
-    border-color: #33C3F0; }
-.button.button-primary:hover,
-button.button-primary:hover,
-input[type="submit"].button-primary:hover,
-input[type="reset"].button-primary:hover,
-input[type="button"].button-primary:hover,
-.button.button-primary:focus,
-button.button-primary:focus,
-input[type="submit"].button-primary:focus,
-input[type="reset"].button-primary:focus,
-input[type="button"].button-primary:focus {
-    color: #FFF;
-    background-color: #1EAEDB;
-    border-color: #1EAEDB; }
-
-.button.button-default,
-button.button-default,
-input[type="submit"].button-default,
-input[type="reset"].button-default,
-input[type="button"].button-default {
-    color: #FFF;
-    background-color: #0775bd;
-    font-weight: bold;
-    border-color: #0C90E6; }
-.button.button-default:hover,
-button.button-default:hover,
-input[type="submit"].button-default:hover,
-input[type="reset"].button-default:hover,
-input[type="button"].button-default:hover,
-.button.button-default:focus,
-button.button-default:focus,
-input[type="submit"].button-default:focus,
-input[type="reset"].button-default:focus,
-input[type="button"].button-default:focus {
-    color: #FFF;
-    background-color: #0C639C;
-    border-color: #0C639C; }
-
-.button.button-reset,
-button.button-reset,
-input[type="submit"].button-reset,
-input[type="reset"].button-reset,
-input[type="button"].button-reset {
-    color: #FFF;
-    background-color: #EA4A4A;
-    font-weight: bold;
-    border-color: #D94444; }
-.button.button-reset:hover,
-button.button-reset:hover,
-input[type="submit"].button-reset:hover,
-input[type="reset"].button-reset:hover,
-input[type="button"].button-reset:hover,
-.button.button-reset:focus,
-button.button-reset:focus,
-input[type="submit"].button-reset:focus,
-input[type="reset"].button-reset:focus,
-input[type="button"].button-reset:focus {
-    color: #FFF;
-    background-color: #CB3A3A;
-    border-color: #B53131; }
-
-
-
