--- public_html/css/catalogo.css
@@ -1,235 +1,0 @@
-
-#catalogo h3,
-#catalogo h4 {
-	font-weight: bold;
-	color:#2E85D2;
-}
-
-#catalogo h3{
-	margin-bottom: 10px;
-    margin-top: 40px;
-}
-#catalogo h4{
-    font-size: 2em;
-    margin-bottom: 10px;
-    margin-top: 10px;
-}
-
-.items {
-	
-}
-.product-card{
-	width: 31%;
-	display: inline-block;
-	padding: 0;
-	margin:0 15px 50px 0;
-}
-.product-card a.product-card-clic{
-	width: 100%;
-	display: block;
-	text-decoration: none;
-}
-.product-card .product-card-image{
-	
-}
-.product-card img.product-card-image{
-
-}
-.product-card .product-card-info {
-	padding-left: 5px;
-	display: block;
-}
-.product-card .product-card-info h5{
-	font-size: 1em;
-	margin: 0;
-	padding: 0;
-	color: #444;
-	font-weight: bold;
-}
-.product-card .product-card-info p{
-	font-size: .7em;
-	margin: 0;
-	padding: 0;
-	color: #444;
-}
-#producto {
-	padding-top: 30px;
-}
-#producto div.seven #producto-desc {
-	padding: 10px 0 10px 0;
-	line-height: 20px;
-	font-size: .8em;
-}
-#producto div.seven #producto-desc p {
-	margin-bottom: 8px;
-}
-
-#producto h3,
-#producto h4 {
-	font-weight: bold;
-	color:#2E85D2;
-}
-
-#producto h3{
-	margin-bottom: 10px;
-    margin-top: 40px;
-}
-#producto h4{
-    font-size: 2em;
-    margin:0;
-    padding: 0;
-}
-#producto .color-img{
-    width: 30px;
-    height: auto;
-}
-/*
-#catalogo h3, h4{
-    font-weight: bold;
-    color:#2E85D2;
-    margin-bottom: 40px;
-    margin-top: 40px;
-}*/
-
-/*
-
-.catItem{
-	border:4px solid #3498db;
-	display: inline-block;
-	vertical-align: middle;
-	width: 40%;
-	height: auto;
-	margin: 20px;
-	padding: 15px;
-}
-
-.catItem a{
-	color:#3498db;
-	vertical-align: middle;
-	text-decoration: none;
-	width: 100%;
-	height: 100%;
-}
-
-.catItem a img{
-	width: 150px;
-	height: 150px;
-	border:2px solid #3498db;
-	border-radius: 10px;
-}
-
-.catItem a span{
-}
-
-.catItem:hover {
-	border:4px solid #e74c3c;
-	box-shadow: 4px 4px 4px #888888;
-}
-
-.catItem:hover > a img{
-	border:2px solid #e74c3c;
-}
-
-.catItem:hover > a span{
-	color:#e74c3c;
-}
-
-.catItemItem{
-	border:4px solid #3498db;
-	display: inline-block;
-	vertical-align: middle;
-	width: 40%;
-	height: auto;
-	margin: 20px;
-	padding-bottom: 15px;
-	padding-top: 15px;
-}
-
-.catItemItem a{
-	color:#3498db;
-	vertical-align: middle;
-	text-decoration: none;
-	width: 100%;
-	height: 100%;
-}
-
-.catItemItem a img{
-	width: 150px;
-	height: 150px;
-	border:2px solid #3498db;
-	border-radius: 10px;
-}
-
-.catItemItem:hover {
-	border:4px solid #e74c3c;
-	box-shadow: 4px 4px 4px #888888;
-}
-
-.catItemItem:hover > a img{
-	border:2px solid #e74c3c;
-}
-
-.catItemItem:hover > a p{
-	color:#e74c3c;
-}
-
-.producto{
-	border:1px solid red;
-}
-
-.sku{
-	font-size: 0.7em;
-}
-
-.mainImage {
-	width: 450px;
-	height: auto;
-	border: 1px solid blue;
-	margin-bottom: 20px;
-}
-
-.color{
-	width: 50px;
-	height: auto;
-}
-*/
-/*.catItemSec{
-	border:4px solid #3498db;
-	display: inline-block;
-	vertical-align: middle;
-	width: 36%;
-	height: auto;
-	margin: 20px;
-	padding-bottom: 15px;
-	padding-top: 15px;
-}
-
-.catItemSec a{
-	color:#3498db;
-	vertical-align: middle;
-	text-decoration: none;
-	width: 100%;
-	height: 100%;
-}
-
-.catItemSec a img{
-	width: 150px;
-	height: 150px;
-	border:2px solid #3498db;
-	border-radius: 10px;
-}
-
-.catItemSec:hover {
-	border:4px solid #e74c3c;
-	box-shadow: 4px 4px 4px #888888;
-}
-
-.catItemSec:hover > a img{
-	border:2px solid #e74c3c;
-}
-
-.catItemSec:hover > a span{
-	color:#e74c3c;
-}
-*/
-
