--- public_html/css/contacto.css
@@ -1,85 +1,0 @@
-.section-contacto{
-	font-family: 'Muli', sans-serif;
-	text-align: left;
-}
-.section-contacto .p {
-	line-height: 1.5em;
-    font-size: 1em;
-    margin-bottom: 15px;
-    color:#3D3D3D;
-}
-.section-contacto h3 {
-	padding-top: 30px; 
-}
-
-.form input[type="text"], .form input[type="email"] {
-	padding: 8px 15px 8px 15px;
-	border: 2px solid #3498db;
-	border-radius: 5px;
-	transition: 0.5s all;
-	-moz-transition: 0.5s all;
-	-webkit-transition: 0.5s all;
-	-o-transition: 0.5s all;
-	width: 95%;
-}
-.form textarea {
-	padding: 10px 0 10px 10px;
-	border: 2px solid #3498db;
-	border-radius: 5px;
-	transition: 0.5s all;
-	-moz-transition: 0.5s all;
-	-webkit-transition: 0.5s all;
-	-o-transition: 0.5s all;
-	width: 100%;
-	resize: none;
-	height: 250px;
-	margin-bottom: 10px;
-}
-.form label {
-	margin-bottom: 15px;
-	font-weight: bold;
-}
-.group-form{
-	margin-bottom: 15px;
-	padding: 0 15px 0 0;
-}
-.group-form label {
-	width: 100%;
-}
-/*.group-form input{
-	width: 95%;
-}*/
-.group-form label, .group-form input{
-	display: block;
-}
-
-h3{
-	font-weight: bold;
-    color:#2E85D2;
-}
-button{
-	margin-top: 10px;
-}
-input[type="text"]:focus, input[type="email"]:focus, textarea:focus{
-	outline: none;
-	border: 2px solid #e74c3c;
-}
-.form .confirmation {
-	background-color: rgba(35,144,232,.2);
-	border-radius: 5px;
-	border: 1px solid rgba(35,144,232,1);
-	margin: 0 0 15px 0;
-	padding: 7px;
-}
-.form .confirmation input[type="checkbox"] {
-	font-size: .9em;
-	font-weight: bold;
-}
-
-.form .confirmation p {
-	margin-top: 5px;
-}
-.note {
-	font-size: .7em;
-	font-style: italic;
-}
