--- public_html/css/mediaqueries.css
@@ -1,63 +1,0 @@
-/* Large Devices, Wide Screens */
-@media only screen and (max-width : 1200px) {
-
-}
-
-/* Medium Devices, Desktops */
-@media only screen and (max-width : 992px) {
-
-}
-
-/* Small Devices, Tablets */
-@media only screen and (max-width : 768px) {
-
-}
-
-/* Extra Small Devices, Phones */ 
-@media only screen and (max-width : 480px) {
-
-    #menu ul{
-        padding: 0;
-        margin: 0;
-        margin-top: 150px;
-    }
-    #menu ul li {
-        display: block;
-        width: 100%;
-    }
-    #menu ul li a{
-        text-align: left;
-        font-size: .8em;
-        width: auto;
-        display: block;
-    }
-
-    #menu ul li a span.it-text, #menu ul li a i{
-        display: inline;
-    }
-    #menu ul li a i {
-        margin-right: 10px;
-    }
-
-    .logo-float{
-        position: relative;
-        top: 50%;
-        left: 35%;
-        border-radius: 100px;
-        width: 123px;
-        height: 120px;
-        text-align: center;
-    }
-    .logo-float img {
-        width: 76px;
-        height: auto;
-        text-align: center;
-        margin-top: 23px;
-    }
-
-}
-
-/* Custom, iPhone Retina */ 
-@media only screen and (max-width : 320px) {
-    
-}
