--- public_html/css/sucursales.css
@@ -1,22 +1,0 @@
-#sucursales h3{
-    font-weight: bold;
-    color:#2E85D2;
-    margin-bottom: 50px;
-}
-#sucursales article {
-	padding-top: 30px;
-	border-top: 1px solid #BCBCBC;
-	margin-bottom: 30px;
-}
-#sucursales article h5 {
-	font-size: 1.5em;
-	color: #2E85D2;
-	margin: 0 0 15px 0;
-	font-weight: bold;
-}
-#sucursales article p {
-	font-size: 0.8em;
-	line-height: 1.8em;
-}
-
-
