+++ public_html/js/function.js
@@ -1,111 +1,1 @@
-$(document).ready(function () {
-    var emailreg = /^[a-zA-Z0-9_\.\-]+@[a-zA-Z0-9\-]+\.[a-zA-Z0-9\-\.]+$/;
-    
-    $(".boton").click(function (){
-        $(".error").remove();
-        
-        if( $(".nombre").val() == "" || !isNaN($(".nombre").val()) ){
-            $(".nombre").focus().after("<span class='error'>Ingrese su nombre</span>");
-            return false;
-        }else if($(".apellido").val()=="" || !isNaN($(".apellido").val()) ){
-        	$(".apellido").focus().after("<span class='error'>Ingrese su apellido</span>");
-        	return false;
-        }else if($(".telefono").val()=="" || isNaN($(".telefono").val()) || $(".telefono").val().length <10){
-        	$(".telefono").focus().after("<span class='error'>Ingrese su telefono de 10 dígitos</span>");
-        	return false;
-        }else if( $(".correo").val() == "" || !emailreg.test($(".correo").val()) ){
-            $(".correo").focus().after("<span class='error'>Ingrese un email correcto</span>");
-            return false;
-        }else if( $(".mensaje").val() == "" ){
-            $(".mensaje").focus().after("<span class='error'>Ingrese un mensaje</span>");
-            return false;
-        }
-    });
-
-    $("#nombre").keyup(function(){
-        if( $(this).val() != "" || !isNaN($(this).val()) ){
-            $(".error").fadeOut();
-            return false;
-        }
-    });
-
-    $(".apellido").keyup(function(){
-        if( $(this).val() != "" || !isNaN($(this).val()) ){
-            $(".error").fadeOut();
-            return false;
-        }
-    });
-
-    $(".telefono").keyup(function(){
-        if($(this).val() !="" || isNaN($(".telefono").val()) || $(".telefono").val().length < 10 ){
-            $(".error").fadeOut();
-            return false;
-        }
-    });
-  
-
-    $(".correo").keyup(function(){
-        if( $(this).val() != "" && emailreg.test($(this).val())){
-            $(".error").fadeOut();
-            return false;
-        }
-    });
-
-});
-
-        function limita(elEvento,maximoCaracteres) {
-          
-          var elemento = document.getElementById("mensaje");
-          
-          var evento = elEvento || window.event;
-          var codigoCaracter = evento.charCode || evento.keyCode;
-       
-          if(codigoCaracter == 37 || codigoCaracter == 39) {
-            return true;
-          }
-      
-          if(codigoCaracter == 8 || codigoCaracter == 46) {
-            return true;
-          }
-          else if(elemento.value.length >= maximoCaracteres ) {
-            return false;
-          }
-          else {
-            return true;
-          }
-        }     
-
-        function actualizaInfo(maximoCaracteres) {
-          
-          var elemento = document.getElementById("mensaje");
-          var info = document.getElementById("info");
-         
-          if(elemento.value.length >= maximoCaracteres ) {
-            info.innerHTML = "(" + maximoCaracteres + ")";
-          }
-          else {
-            info.innerHTML = "(" + (maximoCaracteres-elemento.value.length) + ")";
-          }
-        }
-
-        function limitaT(elEvento,maximoCaracteres) {
-          
-          var elemento = document.getElementById("telefono");
-         
-          var evento = elEvento || window.event;
-          var codigoCaracter = evento.charCode || evento.keyCode;
-      
-          if(codigoCaracter == 37 || codigoCaracter == 39) {
-            return true;
-          }
-
-          if(codigoCaracter == 8 || codigoCaracter == 46) {
-            return true;
-          }
-          else if(elemento.value.length >= maximoCaracteres ) {
-            return false;
-          }
-          else {
-            return true;
-          }
-        } 
+function limita(e,n){var o=document.getElementById("mensaje"),a=e||window.event,r=a.charCode||a.keyCode;return 37==r||39==r?!0:8==r||46==r?!0:o.value.length>=n?!1:!0}function actualizaInfo(e){var n=document.getElementById("mensaje"),o=document.getElementById("info");n.value.length>=e?o.innerHTML="("+e+")":o.innerHTML="("+(e-n.value.length)+")"}function limitaT(e,n){var o=document.getElementById("telefono"),a=e||window.event,r=a.charCode||a.keyCode;return 37==r||39==r?!0:8==r||46==r?!0:o.value.length>=n?!1:!0}$(document).ready(function(){var e=/^[a-zA-Z0-9_\.\-]+@[a-zA-Z0-9\-]+\.[a-zA-Z0-9\-\.]+$/;$(".boton").click(function(){return $(".error").remove(),""!=$(".nombre").val()&&isNaN($(".nombre").val())?""!=$(".apellido").val()&&isNaN($(".apellido").val())?""==$(".telefono").val()||isNaN($(".telefono").val())||$(".telefono").val().length<10?($(".telefono").focus().after("<span class='error'>Ingrese su telefono de 10 dígitos</span>"),!1):""!=$(".correo").val()&&e.test($(".correo").val())?""==$(".mensaje").val()?($(".mensaje").focus().after("<span class='error'>Ingrese un mensaje</span>"),!1):void 0:($(".correo").focus().after("<span class='error'>Ingrese un email correcto</span>"),!1):($(".apellido").focus().after("<span class='error'>Ingrese su apellido</span>"),!1):($(".nombre").focus().after("<span class='error'>Ingrese su nombre</span>"),!1)}),$("#nombre").keyup(function(){return""==$(this).val()&&isNaN($(this).val())?void 0:($(".error").fadeOut(),!1)}),$(".apellido").keyup(function(){return""==$(this).val()&&isNaN($(this).val())?void 0:($(".error").fadeOut(),!1)}),$(".telefono").keyup(function(){return""!=$(this).val()||isNaN($(".telefono").val())||$(".telefono").val().length<10?($(".error").fadeOut(),!1):void 0}),$(".correo").keyup(function(){return""!=$(this).val()&&e.test($(this).val())?($(".error").fadeOut(),!1):void 0})});
