+++ public_html/js/main.js
@@ -1,20 +1,1 @@
-$(document).ready(function () {
-
-$('.tooltip').tooltipster();
-
-});
-
-
-function collapseExpand(idName, icon){
-	var element = document.getElementById(idName);
-	var icon = document.getElementById(icon);
-
-	if ($('#'+idName).hasClass('hidden')) {
-		element.className = "content showing";
-		icon.className = "fa fa-minus-square"
-	}
-	else {
-		element.className = "content hidden";
-		icon.className = "fa fa-plus-square";
-	}
-}
+function collapseExpand(e,a){var n=document.getElementById(e),a=document.getElementById(a);$("#"+e).hasClass("hidden")?(n.className="content showing",a.className="fa fa-minus-square"):(n.className="content hidden",a.className="fa fa-plus-square")}$(document).ready(function(){$(".tooltip").tooltipster()});
